requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
anyio==3.6.1
asgiref==3.7.2
bitvector-for-humans==0.14.1
blynclight==0.12.0
busylight-for-humans==0.20.5
click==8.1.3
colorama==0.4.5
configparser==5.2.0
Django==4.2.6
django-cors-headers==4.3.0
dnspython==2.6.1
fastapi==0.78.0
flake8==4.0.1
Flask==2.1.2
Flask-RESTful==0.3.9
h11==0.13.0
hidapi==0.12.0.post2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
k-means-constrained==0.7.1
loguru==0.5.3
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.22.3
ortools==9.3.10497
pandas==1.4.2
protobuf==3.20.0
pycodestyle==2.8.0
pydantic==1.9.1
pyflakes==2.4.0
pymongo==4.7.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sqlparse==0.4.4
starlette==0.19.1
threadpoolctl==3.1.0
typer==0.4.2
typing_extensions==4.3.0
tzdata==2023.3
uvicorn==0.18.2
vboxapi==1.0
webcolors==1.12
Werkzeug==2.1.2
win32-setctime==1.1.0
